+++ src/types/route.d.ts
@@ -1,8 +1,9 @@
-import { ComponentPublicInstance } from 'vue';
+import type { ComponentPublicInstance } from 'vue';
 
 export type RouteMeta = {
   title: string;
   alternativeTitles?: string[];
+  sortOrder?: number;
   params?: {
     [key: string]: string;
   };